Fix FormationData initialization order, change update frequency
Formation doesn't need to be updated too often. One time each 2 sec should be enough.
Enforced first update after 5 sec so we are sure all members are loaded.
They will start following the master if its alive but will be on start pos if master is not alive for 5 sec before a new master is choosen.